+++ src/qir/runtime/QIR.cpp
@@ -15,8 +15,6 @@
 
 #include <cstddef>
 #include <cstdint>
-#include <sstream>
-#include <string>
 #include <vector>
 
 extern "C" {
@@ -384,37 +382,30 @@ bool __quantum__rt__read_result(Result* result) {
 void __quantum__rt__result_record_output(Result* result, const char* label) {
   const bool bit = __quantum__rt__read_result(result);
   auto& runtime = qir::Runtime::getInstance();
-  runtime.outputValue(bit ? "1" : "0", label);
+  runtime.outputResult(bit, label);
   // Accumulate new measurement bit.
   runtime.appendMeasurementBit(bit);
 }
 
 void __quantum__rt__bool_record_output(bool value, const char* label) {
-  qir::Runtime::getInstance().outputValue(value ? "1" : "0", label);
+  qir::Runtime::getInstance().outputBool(value, label);
 }
 
 void __quantum__rt__int_record_output(int64_t value, const char* label) {
-  qir::Runtime::getInstance().outputValue(std::to_string(value), label);
+  qir::Runtime::getInstance().outputInt(value, label);
 }
 
 void __quantum__rt__float_record_output(double value, const char* label) {
-  // Use std::ostringstream rather than std::to_string.
-  // std::to_string formats with six digits after the decimal point and
-  // can print 0.000000 for very small numbers.
-  // std::ostringstream uses six significant digits by default and
-  // outputs very small numbers with scientific notation.
-  std::ostringstream oss;
-  oss << value;
-  qir::Runtime::getInstance().outputValue(oss.str(), label);
+  qir::Runtime::getInstance().outputFloat(value, label);
 }
 
 void __quantum__rt__tuple_record_output(int64_t elementCount,
                                         const char* label) {
-  qir::Runtime::getInstance().outputContainer(elementCount, label);
+  qir::Runtime::getInstance().outputTuple(elementCount, label);
 }
 
 void __quantum__rt__array_record_output(int64_t size, const char* label) {
-  qir::Runtime::getInstance().outputContainer(size, label);
+  qir::Runtime::getInstance().outputArray(size, label);
 }
 
 } // extern "C"